backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
greenlet==3.1.1

# Data analysis
pandas==2.2.3
numpy==2.1.3
ta==0.11.0

# ML (Phase 4, install when needed)
# scikit-learn==1.6.0
# xgboost==2.1.3
# joblib==1.4.2

# HTTP client
httpx==0.28.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Scheduling
apscheduler==3.10.4

# Config
python-dotenv==1.0.1
pydantic-settings==2.7.0

# Email
aiosmtplib==3.0.2

# Scraping (for Rava, Bonistas, etc.)
beautifulsoup4==4.12.3
lxml==5.3.0

# Yahoo Finance
yfinance==0.2.48

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx  # also used for test client

# Utils
python-multipart==0.0.18
